refactored user into module with creatable
