--- User.ts
@@ -1,18 +1,0 @@
-import { Merchant } from "./Merchant"
-
-export interface User {
-	id: string
-	email: string
-	merchant: Merchant | Merchant[]
-	option: { [key: string]: any }
-}
-
-export namespace User {
-	export function is(value: any | User): value is User {
-		return typeof value == "object" &&
-			typeof value.id == "string" &&
-			typeof value.email == "string" &&
-			(Merchant.is(value.merchant) || Array.isArray(value.merchant) && value.merchant.every(Merchant.is)) &&
-			typeof value.option == "object"
-	}
-}
