+++ User/Creatable.ts
@@ -1,0 +1,30 @@
+import * as gracely from "gracely"
+import { Merchant } from "../Merchant"
+
+export interface Creatable {
+	email: string
+	merchant: Merchant | Merchant[]
+	option: { [key: string]: any }
+}
+
+// tslint:disable: no-shadowed-variable
+export namespace Creatable {
+	export function is(value: Creatable | any): value is Creatable {
+		return typeof value == "object" &&
+			typeof value.email == "string" &&
+			(Merchant.is(value.merchant) || Array.isArray(value.merchant) && value.merchant.every(Merchant.is)) &&
+			typeof value.option == "object"
+	}
+	export function flaw(value: Creatable | any): gracely.Flaw {
+		return {
+			type: "model.User",
+			flaws: typeof value != "object" ? undefined :
+				[
+					typeof value.email == "string" || { property: "email", type: "string" },
+					(Merchant.is(value.merchant) || Array.isArray(value.merchant) && value.merchant.every(Merchant.is)) || { property: "merchant", type: "Merchant | Merchant[]" },
+					typeof value.option == "object" || { property: "option", type: "{ [key: string]: any }" },
+				]
+				.filter(gracely.Flaw.is) as gracely.Flaw[],
+		}
+	}
+}
